Refactor component files/folders more logically
