+++ apps/nowcasting-app/components/charts/gsp-pv-remix-chart/forecast-header-gsp.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { CloseButtonIcon } from "../../icons";
+import { CloseButtonIcon } from "../../icons/icons";
 
 type ForecastHeaderGSPProps = {
   title: string;
