+++ apps/nowcasting-app/components/charts/gsp-pv-remix-chart/index.tsx
@@ -1,10 +1,10 @@
 import { FC } from "react";
 import RemixLine from "../remix-line";
 import useFormatChartData from "../use-format-chart-data";
-import { formatISODateString } from "../../utils";
+import { formatISODateString } from "../../helpers/utils";
 import ForecastHeaderGSP from "./forecast-header-gsp";
 import useGetGspData from "./use-get-gsp-data";
-import Spinner from "../../spinner";
+import Spinner from "../../icons/spinner";
 
 // We want to have the ymax of the graph to be related to the capacity of the GspPvRemixChart
 // If we use the raw values, the graph looks funny, i.e y major ticks are 0 100 232
