+++ apps/nowcasting-app/components/charts/remix-line.tsx
@@ -9,7 +9,10 @@ import {
   ResponsiveContainer,
   Tooltip
 } from "recharts";
-import { convertISODateStringToLondonTime, formatISODateStringHumanNumbersOnly } from "../utils";
+import {
+  convertISODateStringToLondonTime,
+  formatISODateStringHumanNumbersOnly
+} from "../helpers/utils";
 import { theme } from "../../tailwind.config";
 const yellow = theme.extend.colors["ocf-yellow"].DEFAULT;
 export type ChartData = {
