+++ apps/nowcasting-app/components/helpers/globalState.tsx
