+++ apps/nowcasting-app/components/helpers/utils.ts
