+++ apps/nowcasting-app/components/icons/spinner.tsx
