+++ apps/nowcasting-app/components/map/pvLatest.tsx
@@ -9,8 +9,8 @@ import { ActiveUnit, SelectedData } from "./types";
 import { getAllForecastUrl, MAX_POWER_GENERATED } from "../../constant";
 import ButtonGroup from "../../components/button-group";
 import gspShapeData from "../../data/gsp_regions_20220314.json";
-import useGlobalState from "../globalState";
-import { axiosFetcher, formatISODateString, formatISODateStringHuman } from "../utils";
+import useGlobalState from "../helpers/globalState";
+import { axiosFetcher, formatISODateString, formatISODateStringHuman } from "../helpers/utils";
 import { FcAllResData } from "../types";
 import { theme } from "../../tailwind.config";
 import ColorGuideBar from "./color-guide-bar";
