+++ apps/nowcasting-app/pages/index.tsx
@@ -1,11 +1,11 @@
 import { withPageAuthRequired } from "@auth0/nextjs-auth0";
-import Layout from "../components/layout";
+import Layout from "../components/layout/layout";
 import { PvLatestMap } from "../components/map";
 import SideLayout from "../components/side-layout";
 import PvRemixChart from "../components/charts/pv-remix-chart";
 import useAndUpdateSelectedTime from "../components/hooks/use-and-update-selected-time";
 import React, { useState } from "react";
-import SideFooter from "../components/side-layout/side-footer";
+import Header from "../components/layout/header";
 
 export default function Home() {
   useAndUpdateSelectedTime();
@@ -13,7 +13,7 @@ export default function Home() {
   return (
     <Layout>
       <div className="h-full relative pt-16">
-        <SideFooter />
+        <Header />
         <div id="map-container" className="relative float-right h-full" style={{ width: "56%" }}>
           <PvLatestMap />
         </div>
